+++ packages/harness/src/index.ts
@@ -1,0 +1,12 @@
+/** Public surface of @cairn/harness. */
+export * from "./types.js";
+export * from "./interfaces.js";
+export { runHarness } from "./pipeline.js";
+
+export { InlineContextProvider } from "./context/inline.js";
+export { StaticPlanner } from "./planners/static.js";
+export { AssertionCritic } from "./critics/assertion.js";
+export { ConsoleReporter } from "./reporters/console.js";
+export { JsonReporter } from "./reporters/json.js";
+export { FakeDriver } from "./drivers/fake.js";
+export { ChromeDevToolsDriver } from "./drivers/chrome.js";